Convert MethodExecutionInfo to interface
- Convert `MethodExecutionInfo` to interface
- Add `DefaultMethodExecutionInfo` class in `callback` package
